Rewrote atoi and fixed strlen declaration

+++ ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int chars;
+	size_t chars;
 
 	chars = 0;
 	while (str[chars] != '\0')
