+++ ft_strnstr.c
@@ -24,7 +24,7 @@ static int	match(const char *big, const char *little, size_t len, size_t i)
 		i++;
 		j++;
 	}
-	if (j < ft_strlen(little))
+	if (j < (int)ft_strlen(little))
 		return (0);
 	return (1);
 }
